move ui into a new directory and change favicon
